Add files via upload
Features found through the processing of the output from mVista. The percent is the level of similarity between one single example case and every other sequence found via mVista. If a file is labeled xxpercent2, it has been processed once again so that it only contains features that are also at least xx percent similar to each other.